manually rename import paths
