+++ src/electron/go_vpn_tunnel.ts
@@ -23,7 +23,7 @@ import * as errors from '../www/model/errors';
 
 import {ChildProcessHelper} from './process';
 import {RoutingDaemon} from './routing_service';
-import {pathToEmbeddedBinary} from './paths';
+import {pathToEmbeddedBinary} from '../infrastructure/electron/app_paths';
 import {VpnTunnel} from './vpn_tunnel';
 
 const isLinux = platform() === 'linux';
