+++ src/electron/sslibev_badvpn_tunnel.ts
@@ -25,7 +25,7 @@ import * as errors from '../www/model/errors';
 import {isServerReachable} from './connectivity';
 import {ChildProcessHelper} from './process';
 import {RoutingDaemon} from './routing_service';
-import {pathToEmbeddedBinary} from './paths';
+import {pathToEmbeddedBinary} from '../infrastructure/electron/app_paths';
 import {VpnTunnel} from './vpn_tunnel';
 
 const isLinux = platform() === 'linux';
